refactor: remove v3 suffix in file name for pool
