+++ src/pages/Pools/PoolDetail.module.scss
