+++ src/pages/Pools/components/AddLiquidityModal/AddLiquidityModal.tsx
@@ -16,7 +16,7 @@ import { getUsd, toSumDisplay } from 'libs/utils';
 import { estimateShare } from 'pages/Pools/helpers';
 import { useGetPairInfo } from 'pages/Pools/hooks/useGetPairInfo';
 import { useTokenAllowance } from 'pages/Pools/hooks/useTokenAllowance';
-import { useGetPoolDetail } from 'pages/Pools/hookV3';
+import { useGetPoolDetail } from 'pages/Pools/hooks';
 import { FC, useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import {
@@ -201,7 +201,7 @@ export const AddLiquidityModal: FC<ModalProps> = ({ isOpen, close, onLiquidityCh
   };
 
   const handleDepositAndStakeAll = async (amount1: bigint, amount2: bigint) => {
-    if (!pairInfoData) return displayToast(TToastType.TX_FAILED, { message: "Pool information does not exist" });
+    if (!pairInfoData) return displayToast(TToastType.TX_FAILED, { message: 'Pool information does not exist' });
 
     setActionAllLoading(true);
     displayToast(TToastType.TX_BROADCASTING);
