+++ src/pages/Pools/components/StakeLPModal/StakeLPModal.tsx
@@ -10,7 +10,7 @@ import { handleCheckAddress, handleErrorTransaction } from 'helper';
 import useConfigReducer from 'hooks/useConfigReducer';
 import CosmJs from 'libs/cosmjs';
 import { toFixedIfNecessary } from 'pages/Pools/helpers';
-import { useGetPoolDetail, useGetRewardInfoDetail } from 'pages/Pools/hookV3';
+import { useGetPoolDetail, useGetRewardInfoDetail } from 'pages/Pools/hooks';
 import { useGetPairInfo } from 'pages/Pools/hooks/useGetPairInfo';
 import { FC, useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
