+++ src/routes.tsx
@@ -6,7 +6,7 @@ import { Route, Routes } from 'react-router-dom';
 
 import Balance from 'pages/Balance';
 import Pools from 'pages/Pools';
-import PoolDetail from 'pages/Pools/PoolDetailV3';
+import PoolDetail from 'pages/Pools/PoolDetail';
 import UniversalSwap from 'pages/UniversalSwap/index';
 
 export default () => (
